Started building content
I'm working on building content, and I seperated the content into it's
own continer with mucho apps working into it. The first I am building
or rebulding is the name generator.

I am now focusing on puting my name generator into a modal then after
that putting that into a table.

If this works, the character generator will be exactly the same.